refactor: Move PeerInfo, PeerState, and PeerSupports to peer module
